@@ -16,33 +16,33 @@
   "homepage": "https://brynley.github.io",
   "private": true,
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/preset-env": "^7.11.5",
-    "@babel/preset-react": "^7.10.4",
-    "@babel/preset-typescript": "^7.15.0",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-helmet": "^6.1.2",
-    "@types/react-router-dom": "^5.1.9",
-    "@types/styled-components": "^5.1.14",
-    "babel-loader": "^8.1.0",
-    "babel-plugin-module-resolver": "^4.1.0",
-    "babel-plugin-styled-components": "^1.11.1",
-    "copy-webpack-plugin": "^9.0.1",
+    "@babel/core": "^7.25.7",
+    "@babel/preset-env": "^7.25.7",
+    "@babel/preset-react": "^7.25.7",
+    "@babel/preset-typescript": "^7.25.7",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-helmet": "^6.1.11",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/styled-components": "^5.1.34",
+    "babel-loader": "^9.2.1",
+    "babel-plugin-module-resolver": "^5.0.2",
+    "babel-plugin-styled-components": "^2.1.4",
+    "copy-webpack-plugin": "^12.0.2",
     "file-loader": "^6.2.0",
-    "html-loader": "^1.3.1",
-    "html-webpack-plugin": "^5.3.2",
-    "styled-components": "^5.2.0",
-    "typescript": "^4.4.2",
+    "html-loader": "^5.1.0",
+    "html-webpack-plugin": "^5.6.0",
+    "styled-components": "^6.1.13",
+    "typescript": "^5.6.2",
     "url-loader": "^4.1.1",
-    "webpack": "^5.53.0",
-    "webpack-cli": "^4.8.0",
-    "webpack-dev-server": "^4.2.1"
+    "webpack": "^5.95.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.1.0"
   },
   "dependencies": {
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-helmet": "^6.1.0",
-    "react-router-dom": "^5.2.0",
-    "react-svg": "^14.0.13"
+    "react-router-dom": "^6.26.2",
+    "react-svg": "^16.1.34"
   }
 }
