@@ -1,6 +1,6 @@
 import React from 'react'
 import { ThemeProvider } from 'styled-components'
-import { Route, Switch } from 'react-router-dom'
+import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 
 import { brandPink } from './themes'
 
@@ -9,27 +9,31 @@ import Home, { homePageInfo } from './app/pages/Home'
 import About from './app/pages/About'
 import Error from './app/pages/Error'
 
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: (
+      <Base pageTitle={homePageInfo.title} pagePath={homePageInfo.path}>
+        <Home />
+      </Base>
+    ),
+    errorElement: <Error />
+  },
+  {
+    path: '/about',
+    element: (
+      <Base pageTitle={About.pageInfo.title} pagePath={About.pageInfo.path}>
+        <About />
+      </Base>
+    )
+  }
+])
+
 const App = () => {
   return (
     <ThemeProvider theme={brandPink}>
       <brandPink.globalStyles />
-      
-        <Switch>
-          <Route exact path={homePageInfo.path}>
-            <Base pageTitle={homePageInfo.title} pagePath={homePageInfo.path}>
-              <Home />
-            </Base>
-          </Route>
-          <Route path={About.pageInfo.path}>
-            <Base pageTitle={About.pageInfo.title} pagePath={About.pageInfo.path}>
-              <About />
-            </Base>
-          </Route>
-          <Route>
-            <Error />
-          </Route>
-        </Switch>
-      
+      <RouterProvider router={router} />
     </ThemeProvider>
   )
 }
